# frontend/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ZeroAI",
    "version": "0.1.0"
  },
  "paths": {
    "/api/v1/sources": {
      "get": {
        "tags": [
          "sources"
        ],
        "summary": "List Sources",
        "operationId": "listSources",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/NewsSourceRead"
                  },
                  "type": "array",
                  "title": "Response Listsources"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "sources"
        ],
        "summary": "Create Source",
        "operationId": "createSource",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewsSourceCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NewsSourceRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sources/check": {
      "post": {
        "tags": [
          "sources"
        ],
        "summary": "Check Source",
        "description": "Try a feed URL once and say what it is (RSS/Atom), how many items, or why it failed.",
        "operationId": "checkSource",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedCheckRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedCheck"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sources/{source_id}": {
      "patch": {
        "tags": [
          "sources"
        ],
        "summary": "Update Source",
        "operationId": "updateSource",
        "parameters": [
          {
            "name": "source_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Source Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewsSourceUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NewsSourceRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "sources"
        ],
        "summary": "Delete Source",
        "operationId": "deleteSource",
        "parameters": [
          {
            "name": "source_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Source Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/stocks/{ticker}/news": {
      "get": {
        "tags": [
          "stocks"
        ],
        "summary": "Get News",
        "operationId": "getNews",
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticker"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NewsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/stocks/{ticker}/summary": {
      "get": {
        "tags": [
          "stocks"
        ],
        "summary": "Get Summary",
        "description": "Single-shot (non-streaming) summary, handy for scripts and curl.",
        "operationId": "getSummary",
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticker"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StockSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/stocks/{ticker}/summary/stream": {
      "get": {
        "tags": [
          "stocks"
        ],
        "summary": "Stream",
        "description": "SSE: status, news, thinking, summary (partial), then done or error.",
        "operationId": "streamSummary",
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticker"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Server-Sent Events",
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/settings/llm": {
      "get": {
        "tags": [
          "settings"
        ],
        "summary": "Get Llm Settings",
        "operationId": "getLlmSettings",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LLMConfigRead"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "settings"
        ],
        "summary": "Update Llm Settings",
        "operationId": "updateLlmSettings",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LLMConfigUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LLMConfigRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "FeedCheck": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "kind": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "rss",
                  "atom"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Kind"
          },
          "item_count": {
            "type": "integer",
            "title": "Item Count",
            "default": 0
          },
          "sample": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Sample",
            "default": []
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "ok"
        ],
        "title": "FeedCheck",
        "description": "Result of trying a source once."
      },
      "FeedCheckRequest": {
        "properties": {
          "url_template": {
            "type": "string",
            "maxLength": 500,
            "minLength": 8,
            "title": "Url Template"
          },
          "ticker": {
            "type": "string",
            "title": "Ticker",
            "default": "AAPL"
          }
        },
        "type": "object",
        "required": [
          "url_template"
        ],
        "title": "FeedCheckRequest"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "LLMConfigRead": {
        "properties": {
          "provider": {
            "type": "string",
            "enum": [
              "ollama",
              "openai"
            ],
            "title": "Provider"
          },
          "model": {
            "type": "string",
            "title": "Model"
          },
          "base_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Url"
          },
          "thinking": {
            "type": "boolean",
            "title": "Thinking"
          },
          "thinking_effort": {
            "type": "string",
            "enum": [
              "low",
              "medium",
              "high"
            ],
            "title": "Thinking Effort"
          },
          "api_key_set": {
            "type": "boolean",
            "title": "Api Key Set"
          }
        },
        "type": "object",
        "required": [
          "provider",
          "model",
          "base_url",
          "thinking",
          "thinking_effort",
          "api_key_set"
        ],
        "title": "LLMConfigRead"
      },
      "LLMConfigUpdate": {
        "properties": {
          "provider": {
            "type": "string",
            "enum": [
              "ollama",
              "openai"
            ],
            "title": "Provider"
          },
          "model": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Model"
          },
          "base_url": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 300
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Url"
          },
          "thinking": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thinking"
          },
          "thinking_effort": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "low",
                  "medium",
                  "high"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Thinking Effort"
          },
          "api_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Key"
          }
        },
        "type": "object",
        "required": [
          "provider",
          "model"
        ],
        "title": "LLMConfigUpdate"
      },
...
      "NewsItem": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "link": {
            "type": "string",
            "title": "Link"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "published": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published"
          },
          "snippet": {
            "type": "string",
            "title": "Snippet",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "title",
          "link",
          "source"
        ],
        "title": "NewsItem"
      },
      "NewsResponse": {
        "properties": {
          "ticker": {
            "type": "string",
            "title": "Ticker"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/NewsItem"
            },
            "type": "array",
            "title": "Items"
          },
          "errors": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Errors",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "ticker",
          "items"
        ],
        "title": "NewsResponse"
      },
      "NewsSourceCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Name"
          },
          "url_template": {
            "type": "string",
            "maxLength": 500,
            "minLength": 8,
            "title": "Url Template"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "name",
          "url_template"
        ],
        "title": "NewsSourceCreate",
        "description": "Source create payload; the shared URL type validates its feed template."
      },
      "NewsSourceRead": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Name"
          },
          "url_template": {
            "type": "string",
            "maxLength": 500,
            "minLength": 8,
            "title": "Url Template"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "default": true
          },
          "id": {
            "type": "integer",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "name",
          "url_template",
          "id"
        ],
        "title": "NewsSourceRead"
      },
      "NewsSourceUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 80,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "url_template": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500,
                "minLength": 8
              },
              {
                "type": "null"
              }
            ],
            "title": "Url Template"
          },
          "enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enabled"
          }
        },
        "type": "object",
        "title": "NewsSourceUpdate"
      },
      "StockSummary": {
        "properties": {
          "headline": {
            "type": "string",
            "title": "Headline",
            "description": "One sentence, the single most important takeaway."
          },
          "sentiment": {
            "type": "string",
            "enum": [
              "bullish",
              "bearish",
              "neutral",
              "mixed"
            ],
            "title": "Sentiment"
          },
          "key_points": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Key Points",
            "description": "3-6 short bullet points, most important first."
          },
          "risks": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Risks",
            "description": "Risks or negatives mentioned in the news; may be empty."
          },
          "outlook": {
            "type": "string",
            "title": "Outlook",
            "description": "Two sentences max on what traders should watch next."
          }
        },
        "type": "object",
        "required": [
          "headline",
          "sentiment",
          "key_points",
          "outlook"
        ],
        "title": "StockSummary",
        "description": "Structured output the LLM must produce."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}